requirements.txt
httpx
pillow
selenium
markdown
chromedriver-autoinstaller
click
